chore: replace old task
