+++ lib/generate.js
@@ -1,3 +1,4 @@
+const fs = require("fs");
 const path = require("path");
 const merge = require("merge-deep");
 const generate = require("./generate/index");
@@ -12,29 +13,47 @@ const generateSourceFiles = async () => {
 
     const projects = [...graph.getProjects()]; // the project plus all dependencies (the project is at index 0)
     const currentProject = projects[0];
-    const customConfigurations = projects.map(project => project.getCustomConfiguration());
 
     let mergedGenerationConfig = {}; // the "generation" configuration for this library and all dependencies, merged
-    const apiJSONFilesPaths = []; // list of all "api.json" files that must be read
+    const allPackages = [];
 
-    for (const i in customConfigurations) {
-        const customConfig = customConfigurations[i];
-        if (customConfig && customConfig["ui5-tooling-webc"]) {
-            const packages = customConfig["ui5-tooling-webc"].packages || {};
+    for (const i in projects) {
+        const projectName = projects[i].getName(); // f.e. sap.ui.webc.main
+        const projectCustomConfiguration = projects[i].getCustomConfiguration();
+
+        if (projectCustomConfiguration && projectCustomConfiguration["ui5-tooling-webc"]) {
+            const packages = projectCustomConfiguration["ui5-tooling-webc"].packages || {}; // f.e. @ui5/webcomponents
             for (const packageName in packages) {
                 const packageData = packages[packageName];
+
+                // Merge generation config, if any
                 const generationConfig = packageData.generation;
                 if (typeof generationConfig === "object") {
                     mergedGenerationConfig = merge(mergedGenerationConfig, generationConfig || {});
-                    const packagePath = await getPackagePath(packageName);
-                    const apiJSONFilePath = path.join(packagePath, packageData.inputPath, "api.json");
-                    apiJSONFilesPaths.push(apiJSONFilePath);
                 }
+
+                // Manifests
+                let packagePath;
+                try {
+                    packagePath = await getPackagePath(packageName);
+                } catch (e) {
+                    continue; // package not found
+                }
+
+                const manifestFilePath = path.join(packagePath, packageData.inputPath, "custom-elements-internal.json");
+                if (!fs.existsSync(manifestFilePath)) {
+                    continue; // no custom elements manifest for this package (f.e. icons, theming, etc.)
+                }
+                allPackages.unshift({
+                    projectName,
+                    packageName,
+                    manifestFilePath
+                }); // unshift instead of push so that dependencies go first in allPackages
             }
         }
     }
 
-    return generate(currentProject.getName(), mergedGenerationConfig, apiJSONFilesPaths);
+    return generate(currentProject.getName(), mergedGenerationConfig, allPackages);
 };
 
 generateSourceFiles().then(() => {
