+++ lib/generate/builders/ControlFileBuilder.js
@@ -1,11 +1,7 @@
 /* eslint-disable no-tabs */
 const beautify = require('js-beautify').html;
 const reserved = require('reserved-words');
-const getVersion = require("../utils/getVersion.js");
 const beautifyJS = require("../utils/beautifyJS.js");
-
-const ControlModel = require("./ControlModel.js");
-const BuilderBase = require("./BuilderBase.js");
 
 const CUSTOM_CODE_START = `/* CUSTOM CODE START */`;
 const CUSTOM_CODE_END = `/* CUSTOM CODE END */`;
@@ -14,15 +10,10 @@ const quoteReservedWord = str => {
 	return reserved.check(str, 6) ? `"${str}"` : str;
 };
 
-class ControlFileBuilder extends BuilderBase {
-	constructor(config, library, control, allEntities, tagSuffix = undefined, existingContent = "") {
-		super(config, library, allEntities);
+class ControlFileBuilder {
+	constructor(control, existingContent = "") {
 		this.control = control;
-		this.fullName = control.name;
-		this.controlName = control.basename;
-		this.tagSuffix = tagSuffix;
 		this.existingContent = existingContent;
-		this.model = new ControlModel(config, library, control, this.allEntities);
 	}
 
 	formatDescription(description = "", isPrivate = false) {
@@ -32,12 +23,6 @@ class ControlFileBuilder extends BuilderBase {
 		}).join("\n");
 
 		description = description.split(` class="comment-api-title"`).join("");
-
-		// replace tag names with namespaced UI5 controls
-		description = description.replace(/<code>(ui5-.*?)<\/code>/g, (match, contents) => {
-			const control = this.allEntities.find(entity => entity.tagname === contents);
-			return `<code>${control ? control.name : contents}</code>`;
-		});
 
 		// replace links
 		description = description.replace(/<ui5-link.*?href="(.*?)".*?>(.*?)<\/ui5-link>/g, "{@link $1 $2}");
@@ -82,80 +67,76 @@ class ControlFileBuilder extends BuilderBase {
 
 	// The whole control
 	generate() {
-		const version = getVersion(this.config, `${this.library}.${this.controlName}`);
-
 		// Required ones
 		const requiredModules = [
-			`"sap/ui/core/webc/WebComponent"`,
+			`"${this.control.baseClass.path}"`,
 			`"./library"`
 		];
 		const requiredModulesNames = [
-			`WebComponent`,
+			`${this.control.baseClass.name}`,
 			`library`
 		];
 
 		// Label enablement
-		if (this.config.includeLabelEnablement?.includes(this.fullName)) {
+		if (this.control.hasLabelEnablement) {
 			requiredModules.push(`"sap/ui/core/LabelEnablement"`);
 			requiredModulesNames.push(`LabelEnablement`);
 		}
 
 		// Enabled propagator
-		if (this.model.supportsEnabled()) {
+		if (this.control.supportsEnabled) {
 			requiredModules.push(`"sap/ui/core/EnabledPropagator"`);
 			requiredModulesNames.push(`EnabledPropagator`);
 		}
 
 		// Semantic form support
-		if (this.config.includeSemanticFormSupport?.includes(this.fullName)) {
+		if (this.control.hasSemanticFormSupport) {
 			requiredModules.push(`"sap/ui/base/ManagedObjectObserver"`);
 			requiredModulesNames.push(`ManagedObjectObserver`);
 		}
 
 		// Extra libraries
-		[...this.model.extraLibraries].forEach(lib => {
+		[...this.control.libraries].forEach(lib => {
 			requiredModules.push(`"${lib.split(".").join("/")}/library"`); // f.e. sap.ui.core -> "sap/ui/core/library"
 			requiredModulesNames.push(lib.split(".").pop() + "Library"); // f.e. sap.ui.core -> coreLibrary
 		});
 
 		// Form "name" property support
-		if (this.model.hasFormNameProperty()) {
+		if (this.control.hasFormSupport && this.control.properties.find(prop => prop.name === "name")) {
 			let formSupportModulePath = `"./thirdparty/features/InputElementsFormSupport"`;
-			if (this.model.extraLibraries.has("sap.ui.webc.main")) {
+			if (this.control.libraries.includes("sap.ui.webc.main")) {
 				formSupportModulePath = `"sap/ui/webc/main/thirdparty/features/InputElementsFormSupport"`; // for components not in main, use the full path
 			}
-
 			requiredModules.push(formSupportModulePath);
 		}
 
 		// Other features
-		const features = (this.config.features || {})[this.fullName];
-		if (features) {
-			features.forEach(feature => {
+		if (this.control.features) {
+			this.control.features.forEach(feature => {
 				requiredModules.push(`"./thirdparty/features/${feature}"`);
 			});
 		}
 
 		// The third-party component - required
-		requiredModules.push(`"./thirdparty/${this.controlName}"`);
+		requiredModules.push(`"./thirdparty/${this.control.basename}"`);
 
 		let template = `/*!
 		 * \${copyright}
 		 */
 
-		// Provides control ${this.library}.${this.controlName}.
+		// Provides control ${this.control.library}.${this.control.basename}.
 		sap.ui.define([
 			${requiredModules.join(",\n")}
 		], function(${requiredModulesNames.join(", ")}) {
 			"use strict";${this.enumsString}
 
 			/**
-			 * Constructor for a new <code>${this.controlName}</code>.
+			 * Constructor for a new <code>${this.control.basename}</code>.
 			 *
 			 * @param {string} [sId] ID for the new control, generated automatically if no ID is given
 			 * @param {object} [mSettings] Initial settings for the new control
 			 *
-			 * @extends sap.ui.core.webc.WebComponent
+			 * @extends ${this.control.baseClass.fullName}
 			 * @class
 			 *
 			 * ${this.formatDescription(this.control.description)}
@@ -165,29 +146,27 @@ class ControlFileBuilder extends BuilderBase {
 			 *
 			 * @constructor
 			 * @public
-			 * @since ${version}
-			 * @experimental Since ${version} This control is experimental and its API might change significantly.
-			 * @alias ${this.library}.${this.controlName}${this.implementsInterfacesString}
+			 * @since ${this.control.version}
+			 * @experimental Since ${this.control.version} This control is experimental and its API might change significantly.
+			 * @alias ${this.control.name}${this.implementsInterfacesString}
 			 */
-			var ${this.controlName} = WebComponent.extend("${this.library}.${this.controlName}", {
+			var ${this.control.basename} = ${this.control.baseClass.name}.extend("${this.control.name}", {
 				metadata: {
-					library: "${this.library}",
-					tag: "${this.control.tagname}${this.tagSuffix}"
-					${this.interfacesString}${this.propertiesString}${this.aggregationsString}${this.associationsString}${this.eventsString}${this.methodsString}${this.gettersString}${this.designTimeString}
+					library: "${this.control.library}"${this.tagProperty}${this.interfacesString}${this.propertiesString}${this.aggregationsString}${this.associationsString}${this.eventsString}${this.methodsString}${this.gettersString}${this.designTimeString}
 				}
 			});${this.methodsJSDoc}${this.gettersJSDoc}\n`;
 
-			if (this.config.includeLabelEnablement?.includes(this.fullName)) {
-				template += `\nLabelEnablement.enrich(${this.controlName}.prototype);\n`;
+			if (this.control.hasLabelEnablement) {
+				template += `\nLabelEnablement.enrich(${this.control.basename}.prototype);\n`;
 			}
 
-			if (this.model.supportsEnabled()) {
-				template += `\nEnabledPropagator.call(${this.controlName}.prototype);\n`;
+			if (this.control.supportsEnabled) {
+				template += `\nEnabledPropagator.call(${this.control.basename}.prototype);\n`;
 			}
 
 			template += `\n${CUSTOM_CODE_START}${this.getCustomCode()}${CUSTOM_CODE_END}\n`;
 
-			template += `\nreturn ${this.controlName};
+			template += `\nreturn ${this.control.basename};
 		});`;
 
 		template = beautifyJS(template);
@@ -195,8 +174,16 @@ class ControlFileBuilder extends BuilderBase {
 		return template + "\n";
 	};
 
+    get tagProperty() {
+        if (!this.control.tagWithSuffix) {
+            return "";
+        }
+
+        return `,tag: "${this.control.tagWithSuffix}"`;
+    }
+
 	get implementsInterfacesString() {
-		const interfaces = this.model.getInterfaces();
+		const interfaces = this.control.interfaces;
 		if (!interfaces.length) {
 			return "";
 		}
@@ -205,7 +192,7 @@ class ControlFileBuilder extends BuilderBase {
 	}
 
 	get interfacesString() {
-		const interfaces = this.model.getInterfaces();
+		const interfaces = this.control.interfaces;
 		if (!interfaces.length) {
 			return "";
 		}
@@ -216,7 +203,7 @@ class ControlFileBuilder extends BuilderBase {
 	}
 
 	get propertiesString() {
-		const properties = this.model.getProperties();
+		const properties = this.control.properties;
 		if (!properties.length) {
 			return "";
 		}
@@ -247,16 +234,13 @@ class ControlFileBuilder extends BuilderBase {
 				}`;
 			}
 
-			// Extra markup for private properties
-			const visibilityString = prop.private ? `\n,visibility: "hidden"` : "";
-
 			template += `
 
 			/**
 			* ${this.formatDescription(prop.description, prop.private)}
 			*/
 			${quoteReservedWord(prop.name)}: {
-				type: "${type}"${defaultValueString}${mappingString}${visibilityString}
+				type: "${type}"${defaultValueString}${mappingString}
 			},`;
 		});
 
@@ -266,16 +250,29 @@ class ControlFileBuilder extends BuilderBase {
 	};
 
 	get enumsString() {
-		const enums = this.model.getEnums();
+		const enums = this.control.enums;
 		if (!enums.length) {
 			return "";
 		}
 
-		return "\n\n" + enums.map(e => `var ${e.enumVarName} = ${e.enumVarValue};`).join("\n");
+		return "\n\n" + enums.map(e => {
+            const lastDot = e.lastIndexOf(".");
+            const lib = e.substr(0, lastDot);
+            const name = e.substr(lastDot + 1);
+            let libraryVarName;
+
+            if (lib === this.control.library) {
+                libraryVarName = "library";
+            } else {
+                libraryVarName = lib.split(".").pop() + "Library"; // f.e. sap.ui.core -> coreLibrary, sap.ui.webc.main -> mainLibrary
+            }
+
+            return `var ${name} = ${libraryVarName}.${name};`
+        }).join("\n");
 	};
 
 	get aggregationsString() {
-		const aggregations = this.model.getAggregations();
+		const aggregations = this.control.aggregations;
 		if (!aggregations.length) {
 			return "";
 		}
@@ -283,7 +280,7 @@ class ControlFileBuilder extends BuilderBase {
 		let template = "";
 
 		// default aggregation
-		const defaultAggregation = this.model.getDefaultAggregation();
+		const defaultAggregation = this.control.defaultAggregation;
 		if (defaultAggregation) {
 			template += `,\ndefaultAggregation: "${defaultAggregation}"`;
 		}
@@ -313,7 +310,7 @@ class ControlFileBuilder extends BuilderBase {
 	};
 
 	get associationsString() {
-		const associations = this.model.getAssociations();
+		const associations = this.control.associations;
 		if (!associations.length) {
 			return "";
 		}
@@ -351,7 +348,7 @@ class ControlFileBuilder extends BuilderBase {
 	}
 
 	get eventsString() {
-		const events = this.model.getEvents();
+		const events = this.control.events;
 		if (!events.length) {
 			return "";
 		}
@@ -397,7 +394,7 @@ class ControlFileBuilder extends BuilderBase {
 	};
 
 	get methodsString() {
-		const methods = this.model.getMethods();
+		const methods = this.control.methods;
 		if (!methods.length) {
 			return "";
 		}
@@ -409,7 +406,7 @@ class ControlFileBuilder extends BuilderBase {
 	};
 
 	get gettersString() {
-		const getters = this.model.getGetters();
+		const getters = this.control.getters;
 		if (!getters.length) {
 			return "";
 		}
@@ -421,18 +418,16 @@ class ControlFileBuilder extends BuilderBase {
 	}
 
 	get designTimeString() {
-		if (!this.model.hasDesignTime()) {
+		if (!this.control.designtime) {
 			return "";
 		}
 
-		const libraryPath = this.library.split(".").join("/");
-
 		return `,
-		designtime: "${libraryPath}/designtime/${this.controlName}.designtime"`;
+		designtime: "${this.control.designtime}"`;
 	}
 
 	get methodsJSDoc() {
-		const methods = this.model.getMethods();
+		const methods = this.control.methods;
 		if (!methods.length) {
 			return "";
 		}
@@ -452,7 +447,7 @@ class ControlFileBuilder extends BuilderBase {
 
 			result += `
 		* @public
-		* @name ${this.library}.${this.controlName}#${method.name}
+		* @name ${this.control.name}#${method.name}
 		* @function
 		*/`;
 		});
@@ -461,7 +456,7 @@ class ControlFileBuilder extends BuilderBase {
 	};
 
 	get gettersJSDoc() {
-		const getters = this.model.getGetters();
+		const getters = this.control.getters;
 		if (!getters.length) {
 			return "";
 		}
@@ -472,9 +467,9 @@ class ControlFileBuilder extends BuilderBase {
 			result += `
 
 		/**
-		* ${getter.functionDescription}
+		* ${getter.description}
 		* @public
-		* @name ${this.library}.${this.controlName}#${getter.functionName}
+		* @name ${this.control.name}#${getter.functionName}
 		* @function
 		*/`;
 		});
