+++ lib/generate/builders/ControlInstanceBuilder.js
@@ -1,22 +1,17 @@
 /* eslint-disable no-tabs */
-const ControlModel = require("./ControlModel.js");
 const beautifyJS = require("../utils/beautifyJS.js");
-const BuilderBase = require("./BuilderBase.js");
 
-class ControlInstanceBuilder extends BuilderBase {
-	constructor(config, library, control, allEntities, nesting = 1) {
-		super(config, library, allEntities);
-
+class ControlInstanceBuilder {
+	constructor(control, canUseMainButton = false, nesting = 1) {
 		this.control = control;
-		this.className = control.basename;
-		this.model = new ControlModel(config, library, control, allEntities);
+		this.canUseMainButton = canUseMainButton;
 		this.nesting = nesting;
 
 		this.requiredControls = new Set();
 		this.requiredLibraries = new Set();
 
 		this.requiredControls.add(this.control.name);
-		this.requiredLibraries.add(this.library);
+		this.requiredLibraries.add(this.control.library);
 	}
 
 	generate() {
@@ -24,7 +19,7 @@ class ControlInstanceBuilder extends BuilderBase {
 			return "";
 		}
 
-		const template = `new ${this.className}({
+		const template = `new ${this.control.basename}({
 			${this.properties}${this.aggregations}${this.events}
 		})`;
 
@@ -32,7 +27,7 @@ class ControlInstanceBuilder extends BuilderBase {
 	}
 
 	get properties() {
-		const properties = this.model.getProperties();
+		const properties = this.control.mergedProperties;
 		if (!properties.length) {
 			return "";
 		}
@@ -77,17 +72,17 @@ class ControlInstanceBuilder extends BuilderBase {
 	}
 
 	get aggregations() {
-		const aggregations = this.model.getAggregations();
+		const aggregations = this.control.mergedAggregations;
 		if (!aggregations.length) {
 			return "";
 		}
 
 		let template = "";
+		const done = [];
 		aggregations.forEach(agg => {
-			if (agg.multiple) {
+			if (!done.includes(agg.name)) {
 				template += this.multipleAggregation(agg);
-			} else {
-				template += this.singleAggregation(agg);
+				done.push(agg.name);
 			}
 		});
 
@@ -95,7 +90,12 @@ class ControlInstanceBuilder extends BuilderBase {
 	}
 
 	multipleAggregation(agg) {
-		const children = this.generateChildrenForAggregation(agg, 3);
+        let children;
+        if (agg.type === "sap.ui.webc.main.IButton") {
+            children = this.generateSingleButtonForAggregation(agg);
+        } else {
+            children = this.generateChildrenForAggregation(agg, 3);
+        }
 		if (!children) {
 			return "";
 		}
@@ -104,35 +104,18 @@ class ControlInstanceBuilder extends BuilderBase {
 		],`;
 	}
 
-	singleAggregation(agg) {
-		const child = this.generateChildrenForAggregation(agg, 1);
-		if (!child) {
-			return "";
-		}
-		return `${agg.name}: ${child},`;
-	}
-
 	generateChildrenForAggregation(agg, howMany = 1) {
-		if (!this.hasWebcMainButton()) {
+		if (!this.canUseMainButton) {
 			return "";
-		}
-
-		const classOrInterfaceName = agg.type === "sap.ui.core.Control" ? "sap.ui.webc.main.Button" : agg.type;
-		let controls;
-		const entity = this.allEntities.find(entity => entity.name === classOrInterfaceName);
-		if (!entity) {
-			console.log("NOT FOUND", agg, classOrInterfaceName)
-		}
-		if (entity.kind === "class") {
-			controls = [entity];
-		} else {
-			controls = this.allEntities.filter(entity => (entity.implements || []).includes(classOrInterfaceName));
 		}
 
 		const arr = [];
 		for (let i = 0; i < howMany; i++) {
-			const control = controls[i % howMany] || controls[0];
-			const builder = new ControlInstanceBuilder(this.config, this.library, control, this.allEntities, this.nesting + 1);
+			const control = agg.examples[i % howMany] || agg.examples[0];
+			// if (!control) {
+			//	 console.log(agg.name, agg.examples)
+			// }
+			const builder = new ControlInstanceBuilder(control, this.canUseMainButton, this.nesting + 1);
 
 			arr.push(builder.generate());
 
@@ -147,8 +130,31 @@ class ControlInstanceBuilder extends BuilderBase {
 		return arr.filter(str => !!str).join(",\n");
 	}
 
+    generateSingleButtonForAggregation(agg) {
+        if (!this.canUseMainButton) {
+            return "";
+        }
+
+        const arr = [];
+
+        const control = agg.examples.find(ctr => ctr.name === "sap.ui.webc.main.Button");
+        const builder = new ControlInstanceBuilder(control, this.canUseMainButton, this.nesting + 1);
+
+        arr.push(builder.generate());
+
+        [...builder.requiredControls.values()].forEach(rq => {
+            this.requiredControls.add(rq);
+
+            const lastDot = rq.lastIndexOf(".");
+            this.requiredLibraries.add(rq.substr(0, lastDot));
+        });
+
+
+        return arr.filter(str => !!str).join(",\n");
+    }
+
 	get events() {
-		const events = this.model.getEvents();
+		const events = this.control.mergedEvents;
 		if (!events.length) {
 			return "";
 		}
@@ -175,11 +181,11 @@ class ControlInstanceBuilder extends BuilderBase {
 	}
 
 	getButtonsForMethods(instanceVarName, placeAt) {
-		if (!this.hasWebcMainButton()) {
+		if (!this.canUseMainButton) {
 			return "";
 		}
 
-		const methods = [...this.model.getMethods(), ...this.model.getGetters()];
+		const methods = [...this.control.mergedMethods, ...this.control.mergedGetters];
 		if (!methods.length) {
 			return "";
 		}
