+++ lib/generate/builders/KitchenSinkFileBuilder.js
@@ -2,27 +2,25 @@
 const beautifyHTML = require("js-beautify").html;
 
 const ControlInstanceBuilder = require("./ControlInstanceBuilder.js");
-const BuilderBase = require("./BuilderBase.js");
 const beautifyJS = require("../utils/beautifyJS.js");
 
-class KitchenSinkFileBuilder extends BuilderBase {
-	constructor(config, library, allEntities) {
-		super(config, library, allEntities);
+class KitchenSinkFileBuilder {
+	constructor(controls, canUseMainButton) {
 
-		this.controlsWithTestPages = this.libraryEntities.filter(control => !!control.tagname && !this.config?.noControl?.includes(control.name));
+		this.controlsWithTestPages = controls;
 		this.requiredControls = new Set();
 		this.requiredLibraries = new Set();
 		this.instances = [];
 
-		this.controlsWithTestPages.forEach(entry => {
-			const instanceBuilder = new ControlInstanceBuilder(config, library, entry, allEntities);
+		this.controlsWithTestPages.forEach(control => {
+			const instanceBuilder = new ControlInstanceBuilder(control, canUseMainButton);
 
-			const varName = `o${entry.basename}`; // the variable name to assign the instance to
+			const varName = `o${control.basename}`; // the variable name to assign the instance to
 			const instance = instanceBuilder.generate();
-			const methodButtons = instanceBuilder.getButtonsForMethods(varName, `methods${entry.basename}`);
+			const methodButtons = instanceBuilder.getButtonsForMethods(varName, `methods${control.basename}`);
 
 			this.instances.push({
-				entry,
+				control,
 				instance,
 				methodButtons
 			});
@@ -56,10 +54,10 @@ class KitchenSinkFileBuilder extends BuilderBase {
 			<br><br>`).join("\n\n");
 
 		const instancesString = this.instances.map(controlData => {
-			const varName = `o${controlData.entry.basename}`; // the variable name to assign the instance to
+			const varName = `o${controlData.control.basename}`; // the variable name to assign the instance to
 			return `${controlData.methodButtons}
 			 		var ${varName} = ${controlData.instance};
-					${varName}.placeAt("test${controlData.entry.basename}");`;
+					${varName}.placeAt("test${controlData.control.basename}");`;
 		}).join("\n\n");
 
 		const javascript = `
