+++ lib/generate/builders/LibraryFileBuilder.js
@@ -1,34 +1,33 @@
 /* eslint-disable no-tabs */
-const BuilderBase = require("./BuilderBase.js");
-const getVersion = require("../utils/getVersion.js");
 const beautifyJS = require("../utils/beautifyJS.js");
+const path = require("path");
+const fs = require("fs");
 
-class LibraryFileBuilder extends BuilderBase {
-	normalizeName(fullName) {
-		const lastDot = fullName.lastIndexOf(".");
-		const name = fullName.substr(lastDot + 1);
-		return `${this.library}.${name}`;
-	};
+class LibraryFileBuilder {
+    constructor(library, packages) {
+        this.library = library;
+        this.packages = packages;
+    }
 
 	generateLibrary() {
-		const controls = this.libraryEntities.filter(entry => !!entry.tagname && !this.config?.noControl?.includes(entry.name));
-		const types = this.libraryEntities.filter(entry => !entry.abstract && !entry.tagname && !this.config?.noControl?.includes(entry.name) && entry.kind !== "interface");
-		const interfaces = this.libraryEntities.filter(entry => entry.kind === "interface");
+		const controls = this.packages.reduce((acc, p) => { return [...acc, ...p.controls]; }, []);
+		const types = this.packages.reduce((acc, p) => { return [...acc, ...p.enums]; }, []);
+		const interfaces = this.packages.reduce((acc, p) => { return [...acc, ...p.interfaces]; }, []);
+		const designTimeConfig = this.packages.reduce((acc, p) => { return {...acc, ...p.designTimeData}; }, {});
 
-		const typesString = types.map(type => `"${this.normalizeName(type.name)}"`).join(",\n");
-		const interfacesString = interfaces.map(int => `"${this.normalizeName(int.name)}"`).join(",\n");
-		const controlsString = controls.map(control => `"${this.normalizeName(control.name)}"`).join(",\n");
+		const typesString = types.map(type => `"${type.name}"`).join(",\n");
+		const interfacesString = interfaces.map(int => `"${int.name}"`).join(",\n");
+		const controlsString = controls.map(control => `"${control.name}"`).join(",\n");
 
 		let extensionsString = "";
-		const designtimeConfig = this.config.designtime || {};
         const libraryPath = this.library.split(".").join("/");
 		const changeHandlersStrings = [];
-		for (const fullName in designtimeConfig) {
+		for (const fullName in designTimeConfig) {
 			if (!fullName.startsWith(this.library)) {
 				continue;
 			}
 
-			const designtimeData = designtimeConfig[fullName]; // if string -> use dirctly, else build object
+			const designtimeData = designTimeConfig[fullName]; // if string -> use dirctly, else build object
 
 			if (designtimeData === null) {
 				continue; // if undefined, only set the string in the control file, nothing for the library
@@ -70,11 +69,12 @@ class LibraryFileBuilder extends BuilderBase {
 		 */
 		sap.ui.define([
 				"sap/ui/webc/common/library",
+				"sap/ui/base/DataType",
 				"sap/ui/core/Lib",
 				"./thirdparty/Assets",
 				"./library.config",
 			], // library dependency
-			function(commonLibrary, Library) {
+			function(commonLibrary, DataType, Library) {
 
 			"use strict";
 
@@ -91,6 +91,7 @@ class LibraryFileBuilder extends BuilderBase {
 			 */
 			var thisLib = Library.init({
 				name : "${this.library}",
+				apiVersion: 2,
 				version: "\${version}",
 				dependencies : ["sap.ui.core", "sap.ui.webc.common"],
 				noLibraryCSS: true,
@@ -110,31 +111,29 @@ class LibraryFileBuilder extends BuilderBase {
 			});`;
 
 		interfaces.forEach(int => {
-			const version = getVersion(this.config, this.normalizeName(int.name));
 			template += `
 
 			/**
 			 * ${int.description}
 			 *
-			 * @name ${this.normalizeName(int.name)}
+			 * @name ${int.name}
 			 * @interface
-			 * @${int.visibility}
-			 * @since ${version}
-			 * @experimental Since ${version} This API is experimental and might change significantly.
+			 * @public
+			 * @since ${int.version}
+			 * @experimental Since ${int.version} This API is experimental and might change significantly.
 			 */`;
 		});
 
 		types.forEach(type => {
-			const version = getVersion(this.config, `${this.library}.${type.basename}`);
 			template += `
 
 			/**
 			 * ${type.description}
 			 *
 			 * @enum {string}
-			 * @${type.visibility}
-			 * @since ${version}
-			 * @experimental Since ${version} This API is experimental and might change significantly.
+			 * @public
+			 * @since ${type.version}
+			 * @experimental Since ${type.version} This API is experimental and might change significantly.
 			 */
 			thisLib.${type.basename} = {`;
 
@@ -142,14 +141,15 @@ class LibraryFileBuilder extends BuilderBase {
 				template += `
 
 				/**
-				 * ${prop.description || ""}
-				 * @${prop.visibility}
+				 * ${prop.description}
+				 * @public
 				 */
 				${prop.name}: "${prop.name}",`;
 				});
 
 				template += `
 				};
+				DataType.registerEnum("${type.name}", thisLib.${type.basename});
 				`;
 		});
 
@@ -167,6 +167,20 @@ class LibraryFileBuilder extends BuilderBase {
 	generateLibraryConfig() {
 		return `sap.ui.define([], function() {\n\t"use strict";\n});\n`;
 	}
+
+    writeLibraryFiles() {
+        const srcOutputDir = path.join("src/", this.library.split(".").join(path.sep));
+
+        // library.js
+        const libraryFilePath = path.join(srcOutputDir, `library.js`);
+        fs.writeFileSync(libraryFilePath, this.generateLibrary());
+
+        // library.config.js
+        const libraryConfigFilePath = path.join(srcOutputDir, `library.config.js`);
+        if (!fs.existsSync(libraryConfigFilePath)) {
+            fs.writeFileSync(libraryConfigFilePath, this.generateLibraryConfig());
+        }
+    }
 }
 
 module.exports = LibraryFileBuilder;
