+++ lib/generate/builders/TestFilesBuilder.js
@@ -1,36 +1,32 @@
 /* eslint-disable no-tabs */
 const beautifyHTML = require("js-beautify").html;
-
-const BuilderBase = require("./BuilderBase.js");
-const ControlInstanceBuilder = require("./ControlInstanceBuilder.js");
 const beautifyJS = require("../utils/beautifyJS.js");
+const ControlInstanceBuilder = require("./ControlInstanceBuilder");
 
-class TestFilesBuilder extends BuilderBase {
-	constructor(config, library, control, allControls) {
-		super(config, library, allControls);
+class TestFilesBuilder {
+	constructor(control, canUseMainButton) {
 		this.control = control;
-		this.name = this.control.basename;
 
-		const instanceBuilder = new ControlInstanceBuilder(config, this.library, this.control, this.allEntities);
+        const instanceBuilder = new ControlInstanceBuilder(this.control, canUseMainButton, 1);
 
-		this.varName = `o${this.control.basename}`; // the variable name to assign the instance to
+        this.varName = `o${this.control.basename}`; // the variable name to assign the instance to
 
-		// Call these 2 methods first to build all required controls & libraries
-		this.instance = instanceBuilder.generate();
-		this.methodButtons = instanceBuilder.getButtonsForMethods(this.varName, "methodButtons");
+        // Call these 2 methods first to build all required controls & libraries
+        this.instance = instanceBuilder.generate();
+        this.methodButtons = instanceBuilder.getButtonsForMethods(this.varName, "methodButtons");
 
-		this.libraries = instanceBuilder.getLibrariesString();
-		this.requirePaths = instanceBuilder.getRequirePaths();
-		this.requireModules = instanceBuilder.getRequireModules();
+        this.libraries = instanceBuilder.getLibrariesString();
+        this.requirePaths = instanceBuilder.getRequirePaths();
+        this.requireModules = instanceBuilder.getRequireModules();
 	}
 
-	generateTestPage() {
-		const template = `
-		<!DOCTYPE HTML>
+    generateTestPage() {
+        const template = `
+        <!DOCTYPE HTML>
 		<html>
 		<head>
 			<meta charset="utf-8">
-			<title>${this.name} - ${this.library}</title>
+			<title>${this.control.basename} - ${this.control.library}</title>
 			<script src="shared-config.js"></script>
 			<script src="../../../../../resources/sap-ui-core.js"
 				id="sap-ui-bootstrap"
@@ -54,7 +50,7 @@ class TestFilesBuilder extends BuilderBase {
 				});
 			});
 			</script>
-		</head>
+        </head>
 		<body id="body" class="sapUiBody">
 			<div id="methodButtons"></div>
 			<br><br>
@@ -62,10 +58,10 @@ class TestFilesBuilder extends BuilderBase {
 		</body>
 		</html>`;
 
-		return beautifyHTML(template, {
-			wrap_attributes: ["preserve"]
-		});
-	}
+        return beautifyHTML(template, {
+            wrap_attributes: ["preserve"]
+        });
+    }
 
 	generateQUnit() {
 		const instanceNoComments = this.instance.replace(/console\.log/g, "// console.log");
