+++ lib/generate/models/ComponentsPackage.js
@@ -1,5 +1,3 @@
-const normalizeType = require("../../generate/utils/normalizeType");
-
 // String utils
 const kebabToCamelCase = string => toCamelCase(string.split("-"));
 const toCamelCase = parts => {
