--- lib/generate/utils/getVersion.js
@@ -1,5 +1,0 @@
-const getVersion = (config, name) => {
-    return (config.versions || {})[name] || "1.92.0";
-}
-
-module.exports = getVersion;
